refactor(canvas): deepen chat surface interfaces
